+++ src/main/java/com/mineinabyss/looty/ecs/components/Screaming.kt
